improve audit log
